# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/jadwal/{id}": {
            "get": {
                "tags": [
                    "Jadwal"
                ],
                "summary": "Get List jadwal",
                "description": "data jadwal",
                "operationId": "showBerita",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return array model jadwal"
                    }
                }
            }
        },
...
        "/api/datadiri": {
            "post": {
                "tags": [
                    "Pelamar kerja"
                ],
                "summary": "Tambah data pelamar kerja",
                "description": "Endpoint ini digunakan untuk menambah data pelamar kerja.",
                "operationId": "datadiri",
                "requestBody": {
                    "description": "Form tambah pelamar kerja",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "alamat",
                                    "tempat_tanggal_lahir",
                                    "no_hp",
                                    "klasifikasi_pilihan",
                                    "sub_klasifikasi_pilihan"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama pelamar kerja",
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "description": "Alamat pelamar kerja",
                                        "type": "string"
                                    },
                                    "tempat_lahir": {
                                        "description": "Tempat  lahir pelamar kerja",
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "description": "tanggal  lahir pelamar kerja",
                                        "type": "string"
                                    },
                                    "lokasi": {
                                        "description": "lokasi",
                                        "type": "string"
                                    },
                                    "no_hp": {
                                        "description": "Nomor HP pelamar kerja",
                                        "type": "string"
                                    },
                                    "klasifikasi_pilihan": {
                                        "description": "Klasifikasi pilihan pelamar kerja",
                                        "type": "string"
                                    },
                                    "sub_klasifikasi_pilihan": {
                                        "description": "Sub-klasifikasi pilihan pelamar kerja",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "URL foto pelamar kerja",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model user"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/show-profile": {
            "get": {
                "tags": [
                    "Pelamar kerja"
                ],
                "summary": "Get List pelamar kerja",
                "description": "data pelamar kerja",
                "operationId": "showprofile",
                "responses": {
                    "default": {
                        "description": "return array model user"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/update-profile/{id}": {
            "post": {
                "tags": [
                    "Pelamar kerja"
                ],
                "summary": "Update data user",
                "description": "Update data user berdasarkan ID",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data yang akan diperbarui",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "nama": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "tempat_lahir": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "no_hp": {
                                        "type": "string"
                                    },
                                    "klasifikasi_pilihan": {
                                        "type": "string"
                                    },
                                    "sub_klasifikasi_pilihan": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "URL foto pelamar kerja",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "ringkasan_pribadi": {
                                        "type": "string"
                                    },
                                    "pendidikan": {
                                        "type": "string"
                                    },
                                    "riwayat_karir": {
                                        "type": "string"
                                    },
                                    "sertifikat": {
                                        "type": "string"
                                    },
                                    "keahlian": {
                                        "type": "string"
                                    },
                                    "bahasa": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model user"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}